fix: configure custom Prisma Client output paths and adjust imports accordingly. Fix the prisma-client removal when the output directory is the same as the generated client.

+++ src/utils/removeDir.ts
@@ -1,14 +1,20 @@
 import path from 'path';
 import { promises as fs } from 'fs';
 
-export default async function removeDir(dirPath: string, onlyContent: boolean) {
+export default async function removeDir(
+  dirPath: string,
+  onlyContent: boolean,
+  prismaClientPath?: string,
+) {
   const dirEntries = await fs.readdir(dirPath, { withFileTypes: true });
   await Promise.all(
     dirEntries.map(async (dirEntry) => {
-      const fullPath = path.join(dirPath, dirEntry.name);
-      return dirEntry.isDirectory()
-        ? await removeDir(fullPath, false)
-        : await fs.unlink(fullPath);
+      if (path.join(dirEntry.parentPath, dirEntry.name) !== prismaClientPath) {
+        const fullPath = path.join(dirPath, dirEntry.name);
+        return dirEntry.isDirectory()
+          ? await removeDir(fullPath, false)
+          : await fs.unlink(fullPath);
+      }
     }),
   );
   if (!onlyContent) {
